+++ test/vitest/delegated.spec.ts
@@ -59,6 +59,12 @@ describe("Delegated permissions", () => {
       }));
   });
 
+  it("has a string tag", () => {
+    expect(Object.prototype.toString.call(permissions)).toBe(
+      "[object Permissions]",
+    );
+  });
+
   it("cannot be instantiated directly", async () => {
     statusA = await permissions.query(permissionA);